requirements.txt
fastapi
uvicorn
pydantic
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
requests
torch --index-url https://download.pytorch.org/whl/cpu
transformers
python-multipart
python-dotenv
